:hammer: change value from int to long

+++ Server/Src/BazaarOnline.Application/ViewModels/Advertiesements/AdvertiesementPriceDetailViewModel.cs
@@ -5,7 +5,7 @@ namespace BazaarOnline.Application.ViewModels.Advertiesements
 {
     public class AdvertiesementPriceDetailViewModel
     {
-        public int? Value { get; set; }
+        public long? Value { get; set; }
 
         public bool IsAgreement { get; set; }
